+++ packages/stream_chat/CHANGELOG.md
@@ -1,12 +1,18 @@
 ## Upcoming
 
 🛑️ Removed
+
 - The `MessageTranslation` class has been removed. Use the new `i18n` field in the `Message` class instead.
 
 ✅ Added
 
 - The `Message` class now has an `i18n` field for translations
 - The `User` class now has a `language` field for the user's language preference.
+
+🔄 Changed
+
+- `client.user` is now deprecated in favor of `client.currentUser`.
+- `client.userStream` is now deprecated in favor of `client.currentUserStream`.
 
 🐞 Fixed